[refactor] Modularize each function in .js
