+++ web/static/js/osm.js
@@ -1,92 +1,12 @@
-const markerList = [];
-let id = 0;
-let map;
-
-// 自動發送請求設定
-function autoFetchData() {
-    let url = `/update_data?`;
-    fetch(url)
-        .then(response => response.json()) // 將回應 body(type:json) 解析為 promise object
-        .then(data => {
-            while (id < data.length) {
-                const element = data[id];
-                put_icon(map, element['Latitude'], element['Longitude'], element['ImageType'])
-                id++;
-            }
-        })
-        .catch((error) => {
-            console.error('Error:', error); // 如果有錯誤，打印錯誤信息
-        });
-}
-
-// 請求照片設定
-function imgRequest(markerId) {
-    // 構建查詢參數的 URL，傳遞 Marker ID
-    let url = `/submit_data?marker_id=${encodeURIComponent(markerId)}`;
-
-    // 使用 fetch 發送 GET 請求
-    fetch(url)
-        .then(response => response.json()) // 將回應解析為 JSON 格式
-        .then(data => {
-            console.log('Success:', data); // 成功後打印回應數據
-            if (data.image_path) {
-                document.getElementById("target_img").src = data.image_path;
-                document.getElementById("popup_img").src = data.image_path;
-                console.log('Image changed.');
-            }
-        })
-        .catch((error) => {
-            console.error('Error:', error); // 如果有錯誤，打印錯誤信息
-        });
-}
-
-function put_icon(map, latitude, longitude, vehicleType) {
-    // 自定義圖標：汽車、公車、卡車、機車
-    const CarIcon = L.icon({
-        iconUrl: '../static/img/car_icon.png',
-        iconSize: [40, 40],
-    });
-    const TaxiIcon = L.icon({
-        iconUrl: '../static/img/taxi_icon.png',
-        iconSize: [40, 40],
-    });
-    const ScooterIcon = L.icon({
-        iconUrl: '../static/img/scooter_icon.png',
-        iconSize: [40, 40],
-    });
-
-    // 根據訊號類型選擇圖標
-    let selectedIcon;
-    switch (vehicleType) {
-        case 'car':
-            selectedIcon = CarIcon;
-            break;
-        case 'taxi':
-            selectedIcon = TaxiIcon;
-            break;
-        case 'scooter':
-            selectedIcon = ScooterIcon;
-            break;
-        default:
-            selectedIcon = CarIcon; // 默認圖標
-    }
-
-    // 放置標記
-    if (!isNaN(latitude) && !isNaN(longitude)) {
-        const markerId = markerList.length + 1;
-        const marker = L.marker([latitude, longitude], { icon: selectedIcon }).addTo(map);
-        markerList.push({ marker, markerId });
-        // 當標記被點擊時，顯示車輛的資訊
-        marker.on('click', () => {
-            document.getElementById('target_type').value = vehicleType;
-            document.getElementById('target_lng').value = longitude;
-            document.getElementById('target_lat').value = latitude;
-            imgRequest(markerId);
-        });
-    } else {
-        alert('Please enter valid latitude and longitude.');
-    }
-}
+import { put_icon } from './putIcon.js';
+import { autoFetchData } from './pullRequest.js';
+export let map;
+export var counter = {
+    totalCount: 0,
+    carNum: 0,
+    taxiNum: 0,
+    scooterNum: 0
+};
 
 document.addEventListener("DOMContentLoaded", function () {
     // *** 放置地圖
@@ -100,8 +20,9 @@ document.addEventListener("DOMContentLoaded", function () {
         zoomControl: true, // 是否秀出 - + 按鈕
     }).addTo(map);
 
-    interval = setInterval(autoFetchData, 5000);
+    let interval = setInterval(autoFetchData, 5000);
 
+    // 手動新增圖標
     const btnPutMarkers = document.getElementById('put-markers');
     btnPutMarkers.addEventListener('click', e => {
         const latitude = parseFloat(document.getElementById('input_lat').value);
@@ -117,17 +38,17 @@ document.addEventListener("DOMContentLoaded", function () {
     const popupImg = document.getElementById("popup_img");
     const closePopup = document.getElementById("close_popup");
 
-    // 当点击缩略图时，顯示懸浮視窗
+    // 當點擊缩圖時，顯示懸浮視窗
     thumbnailImg.addEventListener("click", function () {
         popup.style.display = "block"; // 顯示懸浮視窗
     });
 
-    // 当点击关闭按钮时，隐藏懸浮視窗
+    // 當點擊關閉按钮时，隐藏懸浮視窗
     closePopup.addEventListener("click", function () {
         popup.style.display = "none"; // 隱藏懸浮視窗
     });
 
-    // 当点击懸浮視窗外部區域时，隱藏窗口
+    // 當點擊懸浮視窗外部區域时，隱藏窗口
     window.addEventListener("click", function (event) {
         if (event.target === popup) {
             popup.style.display = "none";
