+++ web/templates/index.html
@@ -10,9 +10,11 @@
         <link rel="stylesheet" href="../static/css/style.css"/> 
         <link rel="stylesheet" href="https://unpkg.com/leaflet@1.8.0/dist/leaflet.css" />
         
-        <script src="https://unpkg.com/leaflet@1.8.0/dist/leaflet.js"></script>
-        <script src="../static/js/osm.js"></script>
-        <script src="../static/js/counter.js"></script>
+        <script type="module" src="https://unpkg.com/leaflet@1.8.0/dist/leaflet.js"></script>
+        <script type="module" src="../static/js/osm.js"></script>
+        <!-- <script type="module" src="../static/js/counter.js"></script> -->
+        <script type="module" src="../static/js/pullRequest.js"></script>
+        <script type="module" src="../static/js/putIcon.js"></script>
     </head>
     <body>
         <div class="first_title">
